refactor(session): Simplify session management and configuration
- Rename `renew_session_header` to `new_session_header`
- Remove `SessionInfo` struct and directly use `user_id` and `expires_at`
- Move static configurations to `config.rs`
- Streamline session creation logic
