+++ oauth2_passkey/src/session/main/context_token.rs
@@ -9,38 +9,15 @@ use chrono::{Duration, Utc};
 use hmac::{Hmac, Mac};
 use http::{HeaderMap, header::SET_COOKIE};
 use sha2::Sha256;
-use std::{env, sync::LazyLock};
 
 use headers::{Cookie, HeaderMapExt};
 
-use crate::session::{config::USER_CONTEXT_TOKEN_COOKIE, errors::SessionError};
+use crate::session::{
+    config::{AUTH_SERVER_SECRET, USE_CONTEXT_TOKEN_COOKIE, USER_CONTEXT_TOKEN_COOKIE},
+    errors::SessionError,
+};
 
 type HmacSha256 = Hmac<Sha256>;
-
-// We're using a simple string representation for tokens instead of a struct
-// to minimize dependencies and complexity
-
-static AUTH_SERVER_SECRET: LazyLock<Vec<u8>> =
-    LazyLock::new(|| match env::var("AUTH_SERVER_SECRET") {
-        Ok(secret) => secret.into_bytes(),
-        Err(_) => "default_secret_key_change_in_production"
-            .to_string()
-            .into_bytes(),
-    });
-
-static USE_CONTEXT_TOKEN_COOKIE: LazyLock<bool> = LazyLock::new(|| {
-    match env::var("USE_CONTEXT_TOKEN_COOKIE") {
-        Ok(val) => match val.as_str() {
-            "true" => true,
-            "false" => false,
-            _ => panic!(
-                "USE_CONTEXT_TOKEN_COOKIE must be 'true' or 'false', got '{}'.",
-                val
-            ),
-        },
-        Err(_) => true, // Default to true when not specified
-    }
-});
 
 /// Obfuscate user ID to prevent direct exposure
 /// # Arguments
